Create encrypted vocabulary
Write a C program that finds the encrypted form of the words. Program gets the number of words from the user.
• Use a structure array that consists of the word and the length of the word. Use dynamic memory allocation for the 
creation of the said array.
• Write the function simplEncrypt(): That gets the length of a word and add that number to all characters in a word. The 
function should, then, return the encrypted form of the words for them to be displayed at the end of execution at 
main.
Example Run:
Enter the number of words:5
Enter the 1. word : do
Enter the 2. word : not
Enter the 3. word : blank
Enter the 4. word : use
Enter the 5. word : asdfg
The encrypted form of the 1. word is fq.
The encrypted form of the 2. word is qrw.
The encrypted form of the 3. word is gqfsp.
The encrypted form of the 4. word is xvh.
The encrypted form of the 5. word is fxikl.